backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
pandas==1.5.3
scikit-learn==1.2.2
nltk==3.8.1
openai==0.27.8
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.10.1
joblib==1.2.0